+++ rust/operator-binary/src/catalog/config.rs
@@ -13,18 +13,19 @@ use super::{FromTrinoCatalogError, ToCatalogConfig};
 pub struct CatalogConfig {
     /// Name of the catalog
     pub name: String,
+
     /// Properties of the catalog
     pub properties: BTreeMap<String, String>,
+
     /// List of EnvVar that will be added to every Trino container
     pub env_bindings: Vec<EnvVar>,
-    /// Env-Vars that should be exported.
-    /// The value will be read from the file specified.
-    /// You can think of it like `export <key>=$(cat <value>)`
-    pub load_env_from_files: BTreeMap<String, String>,
+
     /// Additional commands that needs to be executed before starting Trino
     pub init_container_extra_start_commands: Vec<String>,
+
     /// Volumes that need to be added to the pod (e.g. for S3 credentials)
     pub volumes: Vec<Volume>,
+
     /// Volume mounts that need to be added to the Trino container (e.g. for S3 credentials)
     pub volume_mounts: Vec<VolumeMount>,
 }
@@ -35,7 +36,6 @@ impl CatalogConfig {
             name: name.into(),
             properties: BTreeMap::new(),
             env_bindings: Vec::new(),
-            load_env_from_files: BTreeMap::new(),
             init_container_extra_start_commands: Vec::new(),
             volumes: Vec::new(),
             volume_mounts: Vec::new(),
@@ -46,17 +46,6 @@ impl CatalogConfig {
 
     pub fn add_property(&mut self, property: impl Into<String>, value: impl Into<String>) {
         self.properties.insert(property.into(), value.into());
-    }
-
-    pub fn add_env_property_from_file(
-        &mut self,
-        property: impl Into<String>,
-        file_name: impl Into<String>,
-    ) {
-        let property = property.into();
-        let env_name = calculate_env_name(&self.name, &property);
-        self.add_property(&property, format!("${{ENV:{env_name}}}"));
-        self.load_env_from_files.insert(env_name, file_name.into());
     }
 
     pub fn add_env_property_from_secret(